Undo revert.
This change was reverted along with the downgrade of super-csv because
it was believed that this change was causing issues with celery task
registration.

It was reverted in https://github.com/edx/edx-platform/pull/25762

The fix turned out to be something different and this and the
super-csv library should both be same to upgrade.  The library
upgrade will come in a follow-on make upgrade with some other
constraints that also need to be lifted.

Actual fix: https://github.com/edx/edx-platform/pull/25822